requirements.txt
fastapi
sentence-transformers
numpy
faiss-cpu
pydantic
uvicorn[standard]